+++ braindecode/models/__init__.py
@@ -7,6 +7,7 @@ from .base import EEGModuleMixin
 from .bendr import BENDR, InterpolatedBENDR
 from .biot import BIOT, InterpolatedBIOT
 from .brainmodule import BrainModule
+from .brant import Brant
 from .cbramod import CBraMod
 from .codebrain import CodeBrain
 from .contrawr import ContraWR
@@ -92,6 +93,7 @@ __all__ = [
     "Deep4Net",
     "DeepSleepNet",
     "BrainModule",
+    "Brant",
     "EEGConformer",
     "EEGDINO",
     "EEGPT",